requirements.txt
asgiref==3.7.2
boto3==1.34.11
botocore==1.34.11
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-bootstrap5==23.4
django-htmx==1.17.2
django-resized==1.0.2
django-storages==1.14.2
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
packaging==23.2
Pillow==10.1.0
pillow-heif==0.14.0
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
urllib3==2.0.7
